requirements.txt
asgiref==3.8.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
